chore: move BindPath code out of runtime package
